fix: consolidate goal invocation tests for pkg-file-count gate
Restructure @you/goal CLI/API parity coverage into existing test files so
pkg/api, pkg/cli, and pkg/cli/run stay within the 15-file package cap.
